% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptiveMinMatchKNNregression.R
\name{adaptiveMinMatchKNNregression}
\alias{adaptiveMinMatchKNNregression}
\title{"Adaptive Minimum Match" KNN regression (AMMKNN)}
\usage{
adaptiveMinMatchKNNregression(trainX, trainY, newX, maxK)
}
\arguments{
\item{trainX}{a data frame containing training independent variables only}

\item{trainY}{a data frame containing training dependent variable only}

\item{newX}{a data frame containing independent variables for new observations}

\item{maxK}{an integer, the maximum number of neighbors to consider}
}
\value{
data.frame object
}
\description{
Runs an "adaptive minimum matching" KNN (AMMKNN) regression procedure and returns the predictions along with the initially provided new/testing data. This returned data frame can also be used to apply a different rule for choosing the value of K, if the user desires. The predicted values of the dependent variable are stored in the returned data.frame in a variable called `MatchDVMeanMin`. This method was developed and proposed by Anshul Kumar, Lisa Walker, and Roger Edwards in the 2021 article "The application of predictive analytics to identify at-risk students in health professions education."
}
\examples{
trainingRowIndex <- sample(1:nrow(mtcars), 0.75*nrow(mtcars))
dtrain <- mtcars[trainingRowIndex, ] # model training data
dtest <- mtcars[-trainingRowIndex, ] # model test data
dtestWithMatches <- adaptiveMinMatchKNNregression(trainX = subset(dtrain, select = -mpg), trainY = dtrain[c("mpg")], newX = subset(dtest, select = -mpg), maxK = 5)
if (!require(Metrics)) install.packages('Metrics')
library(Metrics)
Metrics::rmse(dtestWithMatches$MatchDVMeanMin, dtest$mpg)
plot(dtestWithMatches$MatchDVMeanMin, dtest$mpg, main = "mpg predictions, adaptive minimum match KNN", xlab = "predicted", ylab = "actual")

}
