% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corVarSelect.R
\name{corVarSelect}
\alias{corVarSelect}
\title{Correlation Variable Selection}
\usage{
corVarSelect(mainVarName, df, threshold)
}
\arguments{
\item{mainVarName}{string with name of variable to correlate with all others}

\item{df}{data.frame from which variables under `threshold` will be removed}

\item{threshold}{correlation threshold to use for variable selection}
}
\value{
data.frame object
}
\description{
Identifies which variables in a data.frame (`df`) are correlated with a single selected variable (`mainVarName`) above or at a designated threshold (`threshold`). Returns a new data.frame with only these variables. Returns a data.frame that contains `mainVarName` and all variables in `df` that are correlated above or equal to threshold with `mainVarName`.
}
\examples{
mtcars.mpg.cor.50 <- corVarSelect(mainVarName = "mpg", df = mtcars, threshold = .5)


}
